pyproject.toml
[project]
name = "auto-trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca-backtrader-api>=0.8.0",
    "alpaca-py>=0.43.2",
    "alpaca-trade-api>=0.48",
    "alpha-vantage<3.0.0",
    "backtrader>=1.9.78.123",
    "feedparser>=6.0.12",
    "google-genai>=1.57.0",
    "google-generativeai>=0.8.6",
    "ipython>=9.9.0",
    "langchain>=1.2.3",
    "lumibot>=1.5.5",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "yfinance>=1.0",
]

[dependency-groups]
dev = [
    "pylint>=4.0.4",
    "pytest>=9.0.2",
]
